Exposing RcppTskit_treeseq_xptr in a header for #40 plus polish
