+++ RcppTskit/R/RcppTskit.R
@@ -17,9 +17,9 @@
 #'   to avoid importing the module repeatedly, if it has been imported already
 #'   in which case we use that imported module. Since this process can be
 #'   finicky (it depends on the availability of reticulate Python, module already
-#'   instaled, internet access, etc.)
+#'   installed, internet access, etc.)
 #' @return \code{get_tskit_py} returns \code{tskit} a reticulate Python module
-#'   if succesful or otherwise throws an error (when \code{object_name} exists
+#'   if successful or otherwise throws an error (when \code{object_name} exists
 #'   but is not a reticulate Python module) or returns \code{simpleError}
 #'   (when installation or import failed). \code{check_tskit_py} returns
 #'   \code{TRUE} if \code{object} is a reticulate Python module or \code{FALSE}
@@ -52,7 +52,7 @@ get_tskit_py <- function(object_name = "tskit", force = FALSE) {
     }
   }
 
-  msgSuccess <- paste0('reticulate::py_require("', object_name, '") succeded!')
+  msgSuccess <- paste0('reticulate::py_require("', object_name, '") succeeded!')
   msgFail <- paste0('reticulate::py_require("', object_name, '") failed!')
   e <- simpleError(msgFail)
   if (!reticulate::py_module_available(object_name)) {
