+++ RcppTskit/src/RcppTskit.cpp
@@ -1,21 +1,4 @@
-#include <Rcpp.h>
-#include <tskit.h>
-
-// using namespace Rcpp; // to omit Rcpp:: prefix for whole Rcpp API
-// using Rcpp::IntegerVector; // to omit Rcpp:: prefix for IntegerVector
-
-// Finaliser to free tsk_treeseq_t when it is garbage collected
-// See \url{https://tskit.dev/tskit/docs/stable/c-api.html#c.tsk_treeseq_free}
-// for more details.
-static void RcppTskit_treeseq_xptr_delete(tsk_treeseq_t *ptr) {
-  if (ptr != NULL) {
-    tsk_treeseq_free(ptr);
-    delete ptr;
-  }
-}
-// Define the external pointer type for tsk_treeseq_t with the finaliser
-using RcppTskit_treeseq_xptr = Rcpp::XPtr<tsk_treeseq_t, Rcpp::PreserveStorage,
-                                          RcppTskit_treeseq_xptr_delete, true>;
+#include <RcppTskit.hpp>
 
 //' @title Report the version of installed kastore C API
 //' @details The version is stored in the installed header \code{kastore.h}.
